+++ features/support/app_pages/idv/idv_upload_computer.rb
@@ -1,4 +1,3 @@
-
 require_relative '../base_classes/error_messages.rb'
 
 class IDVComputer < IDmeBase
@@ -59,28 +58,6 @@ class IDVComputer < IDmeBase
         click_link("Verify")
         close_verify_window
       end
-    end
-  end
-
-  def verify_via_kba(action: "none", populate: "none")
-    idv_user = data_for(:experian_user3)
-    fail_idv_user = data_for(:experian_user4)
-
-    case action
-    when "verify", "fail kba"
-      populate = true
-      user = idv_user
-    when "fail experian"
-      populate = true
-      user = fail_idv_user
-    end
-
-    if populate == true
-      populate_fields_kba(data: user)
-      populate_fields_computer(data: user)
-      check_fcra_box
-      click_button("Verify my information")
-      sleep 3
     end
   end